axiom/requirements.txt
smolagents
rich
requests
beautifulsoup4
duckduckgo-search
python-dotenv
lxml
litellm
